Fixed SRAM access

+++ source/hardware/SRAMManager.c
@@ -43,7 +43,7 @@ extern u32 _sram_bss_end;
         /* super's attributes */																		\
         Object_ATTRIBUTES																				\
         /* save space start address */																	\
-        u32* saveSpaceStartAddress;																	    \
+        u16* saveSpaceStartAddress;																	    \
 
 // define the manager
 __CLASS_DEFINITION(SRAMManager, Object);
@@ -70,7 +70,7 @@ static void __attribute__ ((noinline)) SRAMManager_constructor(SRAMManager this)
 
 	__CONSTRUCT_BASE(Object);
 
-    this->saveSpaceStartAddress = (u32*)&_sram_bss_end;
+    this->saveSpaceStartAddress = (u16*)&_sram_bss_end;
 
 	SRAMManager_initialize(this);
 }
@@ -91,7 +91,7 @@ static void SRAMManager_initialize(SRAMManager this)
 	int i = __SRAM_DUMMY_READ_CYCLES;
 	for(; i--;)
 	{
-		u32 dummyChar[__SRAM_DUMMY_READ_LENGTH];
+		u16 dummyChar[__SRAM_DUMMY_READ_LENGTH];
 		SRAMManager_read(this, (BYTE*)&dummyChar, i, sizeof(dummyChar));
 	}
 }
@@ -115,7 +115,7 @@ void SRAMManager_save(SRAMManager this, const BYTE* const source, int memberOffs
 
 	int i = 0;
 
-	u32* destination = this->saveSpaceStartAddress + memberOffset;
+	u16* destination = this->saveSpaceStartAddress + memberOffset;
 	ASSERT(0 == ((int)destination % 2), "SRAMManager::save: odd destination");
 
 	for(; i < dataSize; i++)
@@ -130,7 +130,7 @@ void SRAMManager_read(SRAMManager this, BYTE* destination, int memberOffset, int
 
 	int i = 0;
 
-	u32* source = this->saveSpaceStartAddress + memberOffset;
+	u16* source = this->saveSpaceStartAddress + memberOffset;
 	ASSERT(0 == ((int)source % 2), "SRAMManager::constructor: odd source");
 
 	for(; i < dataSize; i++)
